requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Cors>=4.0.0
PyJWT>=2.7.0
# Para producción con PostgreSQL (descomentar):
# psycopg2-binary>=2.9.0
# gunicorn>=22.0.0